+++ R/FSAsim.R
@@ -5,9 +5,7 @@
 #' This package can be used to simulate data for fisheries stock assessment methods.
 #' 
 #' @import FSA
-#' @import tcltk
 #' @importFrom nlme groupedData nfGroupedData nmGroupedData getGroupsFormula
-#' @importFrom relax gslider slider
 #' @importFrom TeachingDemos cnvrt.coords
 #' @importFrom tidyr spread
 #'