+++ R/srStartsDP.R
@@ -48,18 +48,29 @@
 #' @rdname srStartsDP
 #' @export
 srStartsDP <- function(formula,data=NULL,
-                       type=c("BevertonHolt","Ricker","Shepherd","SailaLorda","independence"),
+                       type=c("BevertonHolt","Ricker","Shepherd",
+                              "SailaLorda","independence"),
                        param=1,min.prop=0.1,max.mult=3,delta.prop=0.005,...) {
   ## some checks
   type <- match.arg(type)
   ## get the R and S vectors
   tmp <- FSA:::iHndlFormula(formula,data,expNumR=1,expNumE=1)
+  if (!tmp$metExpNumR)
+    FSA:::STOP("'formula' must have only one LHS variable.")
+  if (!tmp$Rclass %in% c("numeric","integer"))
+    FSA:::STOP("LHS variable must be numeric.")
+  if (!tmp$metExpNumE)
+    FSA:::STOP("'formula' must have only one RHS variable.")
+  if (!tmp$Eclass %in% c("numeric","integer")) 
+    FSA:::STOP("RHS variable must be numeric.")
   R <- tmp$mf[,tmp$Rname[1]]
   S <- tmp$mf[,tmp$Enames[1]]
   ## find starting values for the type and param
-  sv <- srStarts(formula,data,type,param)
-  if (!requireNamespace("relax")) stop("'srStartsDP' requires the 'relax' package to be installed to construct the dynamic plot.",call.=FALSE)
-  else iSRStartsDynPlot(S,R,type,param,sv,min.prop,max.mult,delta.prop)
+  sv <- FSA::srStarts(formula,data,type,param)
+  if (!iCheckRStudio()) FSA:::STOP("'srStartsDP' only works in RStudio.")
+  if (iChk4Namespace("manipulate")) {
+    iSRStartsDynPlot(S,R,type,param,sv,min.prop,max.mult,delta.prop)
+  }
 }
 
 ##############################################################
@@ -69,42 +80,48 @@ srStartsDP <- function(formula,data=NULL,
 # Dynamics plots for finding starting values -- main function
 #=============================================================
 iSRStartsDynPlot <- function(S,R,type,param,sv,min.prop,max.mult,delta.prop) {
-  ## internal refresh function for the dialog box
-  refresh <- function(...) {
-    p1 <- relax::slider(no=1)
-    if (type!="independence") p2 <- relax::slider(no=2)
-      else p2=NULL
-    if (type %in% c("Shepherd","SailaLorda")) p3 <- relax::slider(no=3)
-      else p3=NULL
-    iSRDynPlot(S,R,type,param,p1,p2,p3)
-  } # end internal refresh
-  
-  ## Main function
-  ## Set the minimum value as a proportion (from min.prop)
-  ## of the starting values, the maximum value at a multiple
-  ## (from max.mult) of the starting values, and the delta
-  ## value at a proportion (from delta.prop) of the starting
-  ## values.  Unlist first to make as a vector.
+  p1 <- p2 <- p3 <- NULL
+  ## Set the minimum value as a proportion (from min.prop) of the starting
+  ## values, the maximum value at a multiple (from max.mult) of the starting
+  ## values, and the delta value at a proportion (from delta.prop) of the
+  ## starting values. Unlist first to make as a vector.
   sl.defaults <- unlist(sv)
   sl.mins <- sl.defaults*min.prop
   sl.maxs <- sl.defaults*max.mult
   sl.deltas <- sl.defaults*delta.prop
   ## Grab names from the sv vector
   sl.names <- names(sl.defaults)
-  ## Make a title
-  sl.ttl <- paste(type,"Stock-Recruit Model")
-  if (type %in% c("BevertonHolt","Ricker")) sl.ttl <- paste0(sl.ttl," #",param)
   ## Set up names that are specific to type and param
-  relax::gslider(refresh,prompt=TRUE,hscale=2,pos.of.panel="left",
-                 title=sl.ttl,sl.names=sl.names,
-                 sl.mins=sl.mins,sl.maxs=sl.maxs,
-                 sl.deltas=sl.deltas,sl.defaults=sl.defaults)
+  if (type=="independence") {  ## just one slider
+    manipulate::manipulate(
+      { iSRDynPlot(S,R,type,param,p1,p2,p3) },
+      p1=manipulate::slider(sl.mins[[1]],sl.maxs[[1]],step=sl.deltas[[1]],
+                            initial=sl.defaults[[1]],label=sl.names[[1]])
+    )
+  } else if (type %in% c("BevertonHolt","Ricker")) {  ## two sliders
+    manipulate::manipulate(
+      { iSRDynPlot(S,R,type,param,p1,p2,p3) },
+      p1=manipulate::slider(sl.mins[[1]],sl.maxs[[1]],step=sl.deltas[[1]],
+                            initial=sl.defaults[[1]],label=sl.names[[1]]),
+      p2=manipulate::slider(sl.mins[[2]],sl.maxs[[2]],step=sl.deltas[[2]],
+                            initial=sl.defaults[[2]],label=sl.names[[2]])
+    )
+  } else {  ## three sliders
+    manipulate::manipulate(
+      { iSRDynPlot(S,R,type,param,p1,p2,p3) },
+      p1=manipulate::slider(sl.mins[[1]],sl.maxs[[1]],step=sl.deltas[[1]],
+                            initial=sl.defaults[[1]],label=sl.names[[1]]),
+      p2=manipulate::slider(sl.mins[[2]],sl.maxs[[2]],step=sl.deltas[[2]],
+                            initial=sl.defaults[[2]],label=sl.names[[2]]),
+      p3=manipulate::slider(sl.mins[[3]],sl.maxs[[3]],step=sl.deltas[[3]],
+                            initial=sl.defaults[[3]],label=sl.names[[3]])
+    )
+  }
 }
 
-#=============================================================
-# Constructs the actual plot in the dynamics plots for finding
-# starting values
-#=============================================================
+#===============================================================================
+# Constructs the actual plot in the dynamics plots for finding starting values
+#===============================================================================
 iSRDynPlot <- function(S,R,type,param,p1,p2=NULL,p3=NULL) {
   ## create a sequence of spawning stock values
   max.S <- max(S,na.rm=TRUE)
@@ -113,7 +130,8 @@ iSRDynPlot <- function(S,R,type,param,p1,p2=NULL,p3=NULL) {
   ## the model type and param
   y <- iSRDynPlot_makeR(x,type,param,p1,p2,p3)
   ## Construct the scatterplot with superimposed model
-  old.par <- graphics::par(mfrow=c(1,2), mar=c(3.5,3.5,1.25,1.25), mgp=c(2,0.4,0), tcl=-0.2, pch=19)
+  old.par <- graphics::par(mfrow=c(1,2),mar=c(3.5,3.5,1.25,1.25),
+                           mgp=c(2,0.4,0),tcl=-0.2,pch=19)
   graphics::plot(S,R,xlab="Parental (Spawner) Stock",ylab="Recruits",
        ylim=c(0,max(R,na.rm=TRUE)),xlim=c(0,max(S,na.rm=TRUE)))
   graphics::lines(x,y,lwd=2,lty=1,col="blue")
@@ -123,11 +141,10 @@ iSRDynPlot <- function(S,R,type,param,p1,p2=NULL,p3=NULL) {
   graphics::par(old.par)
 }
 
-#=============================================================
-# Construct values for R given values of S, a stock-recruit
-# model and parameterization, and values of the parameters.
-# This is called by iSRDynPlot()
-#=============================================================
+#===============================================================================
+# Construct values for R given values of S, a stock-recruit model and
+# parameterization, and values of the parameters. Called by iSRDynPlot()
+#===============================================================================
 iSRDynPlot_makeR <- function(x,type,param,p1,p2,p3=NULL){
   if (type=="BevertonHolt") {
     if (param==1) {